requirements.txt
click>=8.1
flask>=2.3
flask-limiter>=3.7
flask-talisman>=1.1
flask-wtf>=1.2
gunicorn>=21.2
itsdangerous>=2.1
Jinja2>=3.1,<4
line-bot-sdk>=2.4
openai>=1.0.0,<2
Pillow==10.4.0
portalocker>=2.8
pytest>=8.0
python-dotenv>=1.0
redis>=5.0
requests>=2.31
werkzeug>=3.0
